update makefile, update libft with ft_printf, update viewer with ft_printf

+++ libft/ft_putchar.c
@@ -6,13 +6,13 @@
 /*   By: bajeanno <bajeanno@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 14:36:38 by bajeanno          #+#    #+#             */
-/*   Updated: 2022/11/25 13:28:36 by bajeanno         ###   ########lyon.fr   */
+/*   Updated: 2022/12/14 05:12:25 by bajeanno         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_putchar(int c)
+int	ft_putchar(int c)
 {
 	return (write(1, &c, 1));
 }
